refactor(daemon): isolate process supervision
